feat: hyper skill patch as skill only patch
